Suppress clippy::mem_forget
I understand that clippy isn't supposed to pick up on these in user
code, but we've had a couple of instances (often with nightly builds)
where this happens.  Feel free to reject this change, but as I've seen
other projects include clippy warning suppression, I thought I'd ask.

Note that this lint was added in 1.29, so I'm going to say that
suppressing unknown lints isn't needed.